# public/openapi.yaml
openapi: 3.1.0
info:
  title: Agentic Commerce Protocol (ACP)
  description: API for searching products with cashback and checking user rewards.
  version: 1.0.0
servers:
  - url: https://gogocash-acp.web.app/api
    description: Production Server (Next.js Hosting)
paths:
  /searchProducts:
    get:
      operationId: searchProducts
      summary: Search for products
      description: Search the product catalog. Returns products with cashback affiliate links.
      parameters:
        - name: query
          in: query
          description: Search keywords (e.g., "earphone under $50", "running shoes")
          required: true
          schema:
            type: string
        - name: user_email
          in: query
          description: User's email address for cashback tracking
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: Number of results (default 5)
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Successful search
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                  total_results:
                    type: integer
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        product_name:
                          type: string
                        product_price:
                          type: number
                        affiliate_link:
                          type: string
                          description: Clickable link for the user (Contains tracking for cashback)
                        product_card_url:
                          type: string
                          description: Link to view product details page with image
                        image_url:
                          type: string
                          description: Product image URL
                        merchant_name:
                          type: string
                        estimated_cashback:
                          type: number
                        rating:
                          type: number


  /user/profile:
    get:
      operationId: getUserProfile
      summary: Get User Profile
      description: Retrieve user balance, points, and tier.
      parameters:
        - name: user_email
          in: query
          description: User email (Required)
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile found
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      email:
                        type: string
                      balance:
                        type: number
                      go_points:
                        type: number
                      go_tier:
                        type: string

  /user/cashback:
    get:
      operationId: getUserCashback
      summary: Get Cashback History
      description: Retrieve list of past cashback transactions.
      parameters:
        - name: user_email
          in: query
          description: User email (Required)
          required: true
          schema:
            type: string
      responses:
        "200":
          description: History found
          content:
            application/json:
              schema:
                type: object
                properties:
                  cashbacks:
                    type: array
                    items:
                      type: object
                      properties:
                        amount:
                          type: number
                        description:
                          type: string
                        status:
                          type: string
                        createdAt:
                          type: string